Extract joyride things to joyride.ts
