+++ src/nrepl/repl-start.ts
@@ -9,11 +9,9 @@ import * as outputWindow from '../results-output/results-doc';
 import { getConfig } from '../config';
 import * as replSession from './repl-session';
 import * as cljsLib from '../../out/cljs-lib/cljs-lib';
-import { ReplConnectSequence, joyrideDefaults } from './connectSequence';
+import { ReplConnectSequence } from './connectSequence';
 import * as clojureLsp from '../lsp/main';
 import * as calvaConfig from '../config';
-import status from '../status';
-import * as connector from '../connector';
 
 const TEMPLATES_SUB_DIR = 'bundled';
 const DRAM_BASE_URL = 'https://raw.githubusercontent.com/BetterThanTomorrow/dram';
@@ -191,31 +189,6 @@ export async function startStandaloneRepl(
     await eval.loadFile({}, getConfig().prettyPrintingOptions);
     outputWindow.appendPrompt();
   });
-}
-
-export async function prepareForJackingOrConnect() {
-  status.updateNeedReplUi(true);
-  await state.initProjectDir().catch((e) => {
-    void vscode.window.showErrorMessage('Failed initializing project root directory: ', e);
-  });
-  await outputWindow.initResultsDoc();
-  await outputWindow.openResultsDoc();
-  return state.getProjectRootLocal();
-}
-
-export function joyrideJackIn(joyrideExtension: vscode.Extension<any>, projectDir: string) {
-  joyrideExtension.exports
-    .startNReplServer(projectDir)
-    .then((port) => {
-      utilities.setLaunchingState(null);
-      return connector.connect(joyrideDefaults[0], true, 'localhost', port).then(() => {
-        outputWindow.append('; Jack-in done.');
-        outputWindow.appendPrompt();
-      });
-    })
-    .catch((e: Error) => {
-      console.error('Joyride REPL start failed: ', e);
-    });
 }
 
 export function startOrConnectRepl() {
